Merge all of sprint-2 code into master; sprint-2 will be deleted.

+++ TridentTechCourseRegistration/src/main/java/Sezmi/TridentTechCourseRegistration/student/StudentRepository.java
@@ -2,7 +2,6 @@ package Sezmi.TridentTechCourseRegistration.student;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-//the StudentRepository is used to automatically create students in the database using the Spring Data JPA. 
-public interface StudentRepository extends JpaRepository<Student, Integer>{
+public interface StudentRepository extends JpaRepository<Student, Long> {
 
 }
